+++ scripts/data_model_creation/update_data_model.sh
@@ -1,8 +1,9 @@
 #!/usr/bin/env bash
 
 # python ./scripts/join_data_model.py
-python ./scripts/split_model.py
 
-python ./tests/test_generate_manifests.py
+python ./scripts/partition_data_model.py
 
-sh data-dictionary/processes/main_workflow.sh
+python ./tests/test_generate_manifests.py # will test all manifests are generated properly and then update the dca-template-config.json
+
+sh data-dictionary/processes/main_workflow.sh   # update data dictionary with changes